🚀 CI: 배포 스크립트 수정

+++ scripts/deploy.sh
@@ -6,6 +6,8 @@ cd /home/ec2-user/app
 # 환경변수 DOCKER_APP_NAME을 spring으로 설정
 DOCKER_APP_NAME=spring
 
+# slack-web-hook URL 셋팅
+slack_web_hook="https://hooks.slack.com/services/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"
 
 # 실행중인 blue가 있는지 확인
 # 프로젝트의 실행 중인 컨테이너를 확인하고, 해당 컨테이너가 실행 중인지 여부를 EXIST_BLUE 변수에 저장
@@ -27,16 +29,34 @@ if [ -z "$EXIST_BLUE" ]; then
   # 30초 동안 대기
   sleep 30
 
-  # /home/ec2-user/deploy.log: 로그 파일에 "green 중단 시작"이라는 내용을 추가
-  echo "green 중단 시작 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ec2-user/deploy.log
+  BLUE_HEALTH=$(sudo docker-compose -p ${DOCKER_APP_NAME}-blue -f docker-compose.blue.yml ps | grep Up)
 
-  # docker-compose.green.yml 파일을 사용하여 spring-green 프로젝트의 컨테이너를 중지
-  sudo docker-compose -p ${DOCKER_APP_NAME}-green -f docker-compose.green.yml down
+  if [ -z "$BLUE_HEALTH" ]; then
+    # /home/ec2-user/deploy.log: 로그 파일에 "green 중단 시작"이라는 내용을 추가
+      echo "green 중단 시작 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ec2-user/deploy.log
 
-   # 사용하지 않는 이미지 삭제
-  sudo docker image prune -af
+      # docker-compose.green.yml 파일을 사용하여 spring-green 프로젝트의 컨테이너를 중지
+      sudo docker-compose -p ${DOCKER_APP_NAME}-green -f docker-compose.green.yml down
 
-  echo "green 중단 완료 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ec2-user/deploy.log
+       # 사용하지 않는 이미지 삭제
+      sudo docker image prune -af
+
+      echo "green 중단 완료 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ec2-user/deploy.log
+
+  else
+    echo "blue 배포 중 문제 발생 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ec2-user/deploy.log
+    echo "관리자 알람 발송 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ec2-user/deploy.log
+
+    json="{ \"text\": \"blue 배포 중 문제가 발생하여 배포가 비정상 중단되었으니 확인 부탁드립니다 -> 문제 발생 시각: $(date '+%Y-%m-%d %H:%M:%S')\" }"
+
+    echo "json: $json"
+
+
+
+    curl -X POST -H 'Content-type: application/json' --data "$json" "$slack_web_hook"
+
+    echo "관리자 알람 발송완료, 배포 비정상종료 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ec2-user/deploy.log
+  fi
 
 # blue가 실행중이면 green up
 else
@@ -45,14 +65,31 @@ else
 
   sleep 30
 
-  echo "blue 중단 시작 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ec2-user/deploy.log
-  sudo docker-compose -p ${DOCKER_APP_NAME}-blue -f docker-compose.blue.yml down
-  sudo docker image prune -af
+  GREEN_HEALTH=$(sudo docker-compose -p ${DOCKER_APP_NAME}-green -f docker-compose.green.yml ps | grep Up)
 
-  echo "blue 중단 완료 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ec2-user/deploy.log
+  if [ -z "$GREEN_HEALTH" ]; then
+      # /home/ec2-user/deploy.log: 로그 파일에 "blue 중단 시작"이라는 내용을 추가
+        echo "blue 중단 시작 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ec2-user/deploy.log
 
-fi
-  echo "배포 종료  : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ec2-user/deploy.log
+        # docker-compose.blue.yml 파일을 사용하여 spring-green 프로젝트의 컨테이너를 중지
+        sudo docker-compose -p ${DOCKER_APP_NAME}-blue -f docker-compose.blue.yml down
 
-  echo "===================== 배포 완료 =====================" >> /home/ec2-user/deploy.log
-  echo >> /home/ec2-user/deploy.log
+         # 사용하지 않는 이미지 삭제
+        sudo docker image prune -af
+
+        echo "blue 중단 완료 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ec2-user/deploy.log
+
+  else
+      echo "green 배포 중 문제 발생 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ec2-user/deploy.log
+      echo "관리자 알람 발송 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ec2-user/deploy.log
+
+      json="{ \"text\": \"blue 배포 중 문제가 발생하여 배포가 비정상 중단되었으니 확인 부탁드립니다 -> 문제 발생 시각: $(date '+%Y-%m-%d %H:%M:%S')\" }"
+
+      echo "json: $json"
+
+
+
+      curl -X POST -H 'Content-type: application/json' --data "$json" "$slack_web_hook"
+
+      echo "관리자 알람 발송완료, 배포 비정상종료 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ec2-user/deploy.log
+  fi
